requirements.txt
grpcio-tools
requests==2.23.0
transformers
torch
